+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ## New functions
 
-* `get_variances()`, to calculate the variance components from mixed models of class `merMod` and `glmmTMB`.
+* `get_variances()`, to calculate the variance components from mixed models of class `merMod`, `glmmTMB` and `stanreg`.
 
 # insight 0.1.2